rename Prompt to ScanPrompt
