+++ src/components/ScanPrompt.tsx
